requirements.txt
tensorflow-macos
pandas
joblib
Pillow
tqdm
matplotlib
scikit-learn